refactor: Moved fabric-specific data loader to fabric package
