+++ src/main/kotlin/lol/gito/radgyms/fabric/RadGymsDataLoader.kt
@@ -1,17 +1,10 @@
-/*
- * Copyright (c) 2025. gitoido-mc
- * This Source Code Form is subject to the terms of the MIT License.
- * If a copy of the MIT License was not distributed with this file,
- * you can obtain one at https://github.com/gitoido-mc/rad-gyms/blob/main/LICENSE.
- *
- */
-
-package lol.gito.radgyms.common
+package lol.gito.radgyms.fabric
 
 import com.cobblemon.mod.common.util.endsWith
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.decodeFromStream
+import lol.gito.radgyms.common.RadGyms
 import lol.gito.radgyms.common.api.dto.Gym
 import lol.gito.radgyms.common.gym.SpeciesManager
 import lol.gito.radgyms.common.pokecache.CacheDTO
